requirements.txt
asyncpg
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py
git+https://github.com/Gorialis/jishaku@master#egg=jishaku
PyYAML
uvloop
